requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
audioop-lts==0.2.2
av==15.1.0
azure-cognitiveservices-speech==1.46.0
azure-core==1.35.1
bidict==0.23.1
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.0
discord-ext-voice_recv==0.5.2a179
discord.py==2.6.3
diskcache==5.6.3
fastapi==0.117.1
faster-whisper==1.2.0
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
Flask-SocketIO==5.5.1
flatbuffers==25.9.23
frozenlist==1.7.0
fsspec==2025.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.1
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
llama_cpp_python==0.3.16
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.6.4
networkx==3.5
numpy==2.2.6
ollama==0.6.0
onnxruntime==1.23.0
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
propcache==0.3.2
protobuf==6.32.1
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyNaCl==1.5.0
pyreadline3==3.5.4
python-dotenv==1.1.1
python-engineio==4.12.3
python-socketio==5.13.0
pywin32==311
PyYAML==6.0.3
requests==2.32.5
scipy==1.16.2
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.13.1
tokenizers==0.22.1
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
webrtcvad==2.0.10
websockets==15.0.1
Werkzeug==3.1.3
WMI==1.5.1
wsproto==1.2.0
yarl==1.20.1